v1.0 split module && base test support
