+++ demo-api/src/main/java/com/alo7/kelin/ServerApplication.java
@@ -1,0 +1,13 @@
+package com.alo7.kelin;
+
+import org.mybatis.spring.annotation.MapperScan;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+@MapperScan("com.alo7.kelin.mapper")
+public class ServerApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(ServerApplication.class, args);
+    }
+}
