+++ demo-base/src/main/java/com/alo7/kelin/testing/BaseMockTest.java
@@ -1,0 +1,13 @@
+// Copyright 2019 Alo7 Inc. All rights reserved.
+
+package com.alo7.kelin.testing;
+
+import org.junit.runner.RunWith;
+import org.mockito.junit.MockitoJUnitRunner;
+
+/**
+ * @author Kelin Tan
+ */
+@RunWith(MockitoJUnitRunner.class)
+public class BaseMockTest extends BaseTest {
+}
