+++ demo-base/src/main/java/com/alo7/kelin/testing/BaseMockWebTest.java
@@ -1,0 +1,35 @@
+// Copyright 2019 Alo7 Inc. All rights reserved.
+
+package com.alo7.kelin.testing;
+
+import org.junit.Before;
+import org.junit.runner.RunWith;
+import org.mockito.MockitoAnnotations;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+
+/**
+ *
+ * @author Kelin Tan
+ */
+@RunWith(SpringJUnit4ClassRunner.class)
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
+public class BaseMockWebTest extends BaseTest {
+    @Autowired
+    protected WebApplicationContext webApplicationContext;
+
+    protected MockMvc mockMvc;
+
+    @Before
+    public void setUp() {
+        //注入mockMvc用于rest api测试
+        mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext).build();
+        //初始化mock注解支持
+        MockitoAnnotations.initMocks(this);
+    }
+}
