+++ demo-common/src/main/resources/mapper/UserMapper.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!-- Copyright 2018 Alo7 Inc. All rights reserved. -->
+
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
+<mapper namespace="com.alo7.kelin.mapper.UserMapper">
+
+    <sql id="columns">
+        id,user_name
+    </sql>
+
+    <insert id="insert" parameterType="User" useGeneratedKeys="true" keyProperty="id">
+        INSERT INTO
+        t_user(user_name)
+        VALUES(#{userName})
+    </insert>
+
+    <select id="findAll" resultType="User">
+        SELECT
+        <include refid="columns"/>
+        FROM
+        t_user
+    </select>
+
+    <select id="findById" resultType="User">
+        SELECT
+        <include refid="columns"/>
+        FROM
+        t_user
+        where
+        id = #{id}
+    </select>
+</mapper>
