--- src/main/java/com/jrs/kelin/SpringBootServerApplication.java
@@ -1,13 +1,0 @@
-package com.jrs.kelin;
-
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-@MapperScan("com.jrs.kelin.mapper")
-public class SpringBootServerApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(SpringBootServerApplication.class, args);
-    }
-}
